+++ src/platform.ts
@@ -24,6 +24,8 @@ import { Subscription } from "rxjs";
 import { RoomTemperatureControllerAccessory } from "./roomTemperatureControllerAccessory";
 import { EmptyGuid } from "./util";
 
+const DelayFactor = 200;
+
 /** The free&#64;home Homebridge platform. */
 export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
   /** The service reference */
@@ -35,9 +37,11 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
   public readonly accessories: Array<PlatformAccessory<FreeAtHomeContext>> = [];
   /** The system access point */
   public readonly sysap: SystemAccessPoint;
-  private fahAccessories = new Map<string, FreeAtHomeAccessory>();
-  private fahLogger: FreeAtHomeLogger;
+  private readonly fahAccessories = new Map<string, FreeAtHomeAccessory>();
+  private readonly fahLogger: FreeAtHomeLogger;
   private readonly webSocketSubscription: Subscription;
+  private wsConnectionAttempt = 0;
+  private readonly maxWsRetryCount: number;
 
   /**
    * Constructs a new free&#64;home Homebridge platform instance.
@@ -50,16 +54,19 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
     public readonly config: PlatformConfig,
     public readonly api: API
   ) {
+    // set maximum reconnection attempt count
+    this.maxWsRetryCount = (this.config.maxWsRetryCount as number) ?? 10;
+
     // Create a logger for the free&#64;home Local API Client
     this.fahLogger = {
       debug: (message?: unknown, ...optionalParams: unknown[]) =>
-        log.debug((message as string) ?? "", ...optionalParams),
+        log.debug(<string>message, ...optionalParams),
       error: (message?: unknown, ...optionalParams: unknown[]) =>
-        log.error((message as string) ?? "", ...optionalParams),
+        log.error(<string>message, ...optionalParams),
       log: (message?: unknown, ...optionalParams: unknown[]) =>
-        log.info((message as string) ?? "", ...optionalParams),
+        log.info(<string>message, ...optionalParams),
       warn: (message?: unknown, ...optionalParams: unknown[]) =>
-        log.warn((message as string) ?? "", ...optionalParams),
+        log.warn(<string>message, ...optionalParams),
     };
 
     // Create a system access point instance
@@ -71,6 +78,33 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
       this.config.verboseErrors as boolean,
       this.fahLogger
     );
+
+    // React to web socket events
+    this.sysap.on("websocket-open", () => {
+      this.wsConnectionAttempt = 0;
+    });
+    this.sysap.on("websocket-close", (code: number, reason: Buffer) => {
+      if (code === 1000) return;
+
+      this.log.warn(
+        `Websocket to System Access Point was closed with code ${code.toString()}: ${reason.toString()}`
+      );
+      if (this.wsConnectionAttempt >= this.maxWsRetryCount) {
+        this.log.error(
+          "Maximum retry count exceeded. Will not try to reconnect to websocket again."
+        );
+        return;
+      }
+
+      const delay = DelayFactor * 2 ** this.wsConnectionAttempt++;
+      this.log.info(
+        `Attempting to reconnect in ${delay}ms [${this.wsConnectionAttempt}/${this.maxWsRetryCount}]`
+      );
+      setTimeout(
+        () => this.sysap.connectWebSocket(this.config.tlsEnabled as boolean),
+        delay
+      );
+    });
 
     // Subscribe to web socket messages
     this.webSocketSubscription = this.sysap
@@ -137,7 +171,9 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
 
       // Enumerate the channels
       Object.keys(device.channels).forEach((channelId: string) => {
-        const channel = device.channels?.[channelId];
+        // We are enumerating the keys of the channels object. Neither the channels object nor the channelId can possibly be undefined.
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+        const channel = device.channels![channelId];
         if (
           !this.isViableChannel(
             serial,
@@ -186,7 +222,10 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
           ]);
         }
 
-        // Ignore non free&#64;home accessories
+        // This check is used to apply the type guard so the accessory can be used as a free@home accesory without a cast.
+        // Given that the accessory context is constructed in the previous lines, it is impossible for the type check to fail.
+        // Consequently the branch can never be covered and is excluded from the coverage.
+        /* istanbul ignore next */
         if (!isFreeAtHomeAccessory(accessory, this.fahLogger)) return;
 
         // create accessory
@@ -195,19 +234,20 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
       });
     });
   }
-  isViableChannel(
+
+  private isViableChannel(
     serial: string,
     channelId: string,
-    channel: Channel | undefined,
+    channel: Channel,
     locationConfiguredOnDeviceLevel: boolean
   ): channel is Channel {
-    if (!channel) return false;
-
     // Filter unsupported channels
     if (
       !(
         channel.functionID &&
-        Object.values<string>(FunctionID).includes(channel.functionID)
+        Object.values<string>(FunctionID).includes(
+          channel.functionID.toUpperCase()
+        )
       )
     ) {
       this.log.debug(
@@ -233,9 +273,10 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
     functionID: string,
     channelId: string,
     accessory: PlatformAccessory<FreeAtHomeContext>
-  ) {
-    switch (functionID) {
-      case FunctionID.FID_SWITCH_ACTUATOR:
+  ): void {
+    switch (functionID.toUpperCase()) {
+      case FunctionID.FID_DES_AUTOMATIC_DOOR_OPENER_ACTUATOR:
+        // case FunctionID.FID_SWITCH_ACTUATOR:
         this.fahAccessories.set(
           `${serial}_${channelId}`,
           new SwitchActuatorAccessory(this, accessory)
@@ -276,13 +317,9 @@ export class FreeAtHomeHomebridgePlatform implements DynamicPlatformPlugin {
         return;
       }
 
-      // Ignore data points we don't have an accessory for
-      const identifier = `${match[1]}_${match[2]}`;
-      if (!this.fahAccessories.has(identifier)) return;
-
-      // Update the accessory data point
+      // Ignore the data point if we don't have an accessory for it or update the accessory
       this.fahAccessories
-        .get(identifier)
+        .get(`${match[1]}_${match[2]}`)
         ?.updateDatapoint(match[3], message[EmptyGuid].datapoints[datapoint]);
     });
   }