+++ src/util.ts
@@ -1,5 +1,2 @@
 // In the local API the system access point UUID is always an empty UUID. Could be extended later to also support the cloud API.
 export const EmptyGuid = "00000000-0000-0000-0000-000000000000";
-
-/** The default debounce timer  */
-export const DefaultDebounce = 250;