+++ CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0-pre - 20.06.2022
+
+### Added
+
+- [PG-186](https://pgerke.atlassian.net/browse/PG-186):
+  Automatic reconnection to System Access Point web socket after connection loss.
+- [PG-192](https://pgerke.atlassian.net/browse/PG-192):
+  Support automatic door opener
+- [PG-190](https://pgerke.atlassian.net/browse/PG-190):
+  Meet test coverage target
+
+### Fixed
+
+- [PG-193](https://pgerke.atlassian.net/browse/PG-193):
+  Rolled back to ts-node@10.8.0 to fix an issue with code coverage calculation.
+
+## 0.2.3-pre - 17.06.2022
+
+### Fixed
+
+- [PG-188](https://pgerke.atlassian.net/browse/PG-188), [#7](https://github.com/pgerke/homebridge-freeathome-local-api/issues/7):
+  Fixed dimmer brightness fallback issue
+
 ## 0.2.2-pre - 16.06.2022
 
 ### Fixed